+++ components/visualizers/ThreeScenes.tsx
@@ -1,9 +1,9 @@
 
 /**
  * File: components/visualizers/ThreeScenes.tsx
- * Version: 1.4.0
+ * Version: 1.5.0
  * Author: Aura Vision Team
- * Updated: 2025-03-14 23:20
+ * Updated: 2025-03-16 10:15
  */
 
 export { KineticWallScene } from './scenes/KineticWallScene';
@@ -12,3 +12,4 @@ export { CubeFieldScene } from './scenes/CubeFieldScene';
 export { NeuralFlowScene } from './scenes/NeuralFlowScene';
 export { DigitalGridScene } from './scenes/DigitalGridScene';
 export { SilkWaveScene } from './scenes/SilkWaveScene';
+export { OceanWaveScene } from './scenes/OceanWaveScene';